third_party/blink/web_tests/fast/forms/number/number-input-changeevent-expected.txt

This test ensures change event is fired for input[type=number] when the user reverts the change made by a script.

PASS eventSender.keyDown('1');eventSender.keyDown('2');dummy.focus();numberOfChangeEventsFired is 1
PASS test.value = '34';numberOfChangeEventsFired is 1
PASS test.focus();test.value = '34';document.execCommand('SelectAll');
eventSender.keyDown('1');eventSender.keyDown('2');dummy.focus();numberOfChangeEventsFired is 2
PASS successfullyParsed is true

TEST COMPLETE
